# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Middle East Hustle API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PollOption {
  id: number;
  text: string;
  voteCount: number;
  percentage: number;
}

export type PollPollType = (typeof PollPollType)[keyof typeof PollPollType];

export const PollPollType = {
  binary: "binary",
  multiple_choice: "multiple_choice",
  head_to_head: "head_to_head",
  hot_take: "hot_take",
} as const;

export interface Poll {
  id: number;
  question: string;
  context?: string;
  category: string;
  categorySlug: string;
  tags: string[];
  pollType: PollPollType;
  options: PollOption[];
  totalVotes: number;
  isFeatured: boolean;
  isEditorsPick: boolean;
  createdAt: string;
  endsAt?: string | null;
  relatedProfileIds: number[];
}

export interface PollListResponse {
  polls: Poll[];
  total: number;
}

export interface VoteRequest {
  optionId: number;
  voterToken: string;
}

export interface VoteResult {
  success: boolean;
  options: PollOption[];
  totalVotes: number;
  alreadyVoted: boolean;
}

export interface Profile {
  id: number;
  name: string;
  headline: string;
  role: string;
  company?: string;
  sector: string;
  country: string;
  city: string;
  imageUrl?: string | null;
  isFeatured: boolean;
  isVerified: boolean;
  viewCount: number;
  associatedPollCount: number;
}

export type ProfileDetail = Profile & {
  summary: string;
  story: string;
  lessonsLearned: string[];
  quote: string;
  relatedPolls: Poll[];
  similarProfiles: Profile[];
};

export interface ProfileListResponse {
  profiles: Profile[];
  total: number;
}

export interface RankingEntry {
  rank: number;
  label: string;
  sublabel?: string | null;
  score: number;
  /** Position change from last period (+/-) */
  change: number;
  imageUrl?: string | null;
}

export interface RankingsResponse {
  type: string;
  title: string;
  entries: RankingEntry[];
}

export interface Category {
  slug: string;
  name: string;
  pollCount: number;
  icon: string;
}

export interface CategoryListResponse {
  categories: Category[];
}

export type WeeklyPulseStatSentiment =
  (typeof WeeklyPulseStatSentiment)[keyof typeof WeeklyPulseStatSentiment];

export const WeeklyPulseStatSentiment = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;

export interface WeeklyPulseStat {
  label: string;
  value: string;
  change?: string | null;
  sentiment: WeeklyPulseStatSentiment;
}

export type WeeklyPulseSectorSentimentItem = {
  sector: string;
  sentiment: number;
  label: string;
};

export interface WeeklyPulse {
  weekLabel: string;
  topPolls: Poll[];
  biggestSurprise: string;
  stats: WeeklyPulseStat[];
  sectorSentiment: WeeklyPulseSectorSentimentItem[];
}

export type ListPollsParams = {
  filter?: ListPollsFilter;
  category?: string;
  limit?: number;
  offset?: number;
};

export type ListPollsFilter =
  (typeof ListPollsFilter)[keyof typeof ListPollsFilter];

export const ListPollsFilter = {
  latest: "latest",
  trending: "trending",
  most_voted: "most_voted",
  most_debated: "most_debated",
  ending_soon: "ending_soon",
  editors_picks: "editors_picks",
} as const;

export type ListProfilesParams = {
  search?: string;
  country?: string;
  sector?: string;
  filter?: ListProfilesFilter;
  limit?: number;
  offset?: number;
};

export type ListProfilesFilter =
  (typeof ListProfilesFilter)[keyof typeof ListProfilesFilter];

export const ListProfilesFilter = {
  featured: "featured",
  newest: "newest",
  most_viewed: "most_viewed",
  most_followed: "most_followed",
} as const;

export type GetRankingsParams = {
  type?: GetRankingsType;
};

export type GetRankingsType =
  (typeof GetRankingsType)[keyof typeof GetRankingsType];

export const GetRankingsType = {
  founders: "founders",
  women_leaders: "women_leaders",
  sectors: "sectors",
  cities: "cities",
  profiles: "profiles",
  topics: "topics",
} as const;