# lib/api-zod/src/generated/types/poll.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Middle East Hustle API
 * OpenAPI spec version: 0.1.0
 */
import type { PollOption } from "./pollOption";
import type { PollPollType } from "./pollPollType";

export interface Poll {
  id: number;
  question: string;
  context?: string;
  category: string;
  categorySlug: string;
  tags: string[];
  pollType: PollPollType;
  options: PollOption[];
  totalVotes: number;
  isFeatured: boolean;
  isEditorsPick: boolean;
  createdAt: string;
  endsAt?: string | null;
  relatedProfileIds: number[];
}