# lib/api-zod/src/generated/types/pollOption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Middle East Hustle API
 * OpenAPI spec version: 0.1.0
 */

export interface PollOption {
  id: number;
  text: string;
  voteCount: number;
  percentage: number;
}