# lib/api-zod/src/generated/types/profileDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Middle East Hustle API
 * OpenAPI spec version: 0.1.0
 */
import type { Poll } from "./poll";
import type { Profile } from "./profile";

export type ProfileDetail = Profile & {
  summary: string;
  story: string;
  lessonsLearned: string[];
  quote: string;
  relatedPolls: Poll[];
  similarProfiles: Profile[];
};