# lib/api-zod/src/generated/types/profileListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Middle East Hustle API
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from "./profile";

export interface ProfileListResponse {
  profiles: Profile[];
  total: number;
}