# lib/api-zod/src/generated/types/rankingsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Middle East Hustle API
 * OpenAPI spec version: 0.1.0
 */
import type { RankingEntry } from "./rankingEntry";

export interface RankingsResponse {
  type: string;
  title: string;
  entries: RankingEntry[];
}